requirements.txt
numpy
pyqt5
pyopengl
pyopengl_accelerate
pywavefront